# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Coffee API",
    "description": "Allows users get and create coffee orders"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "tags": [
    {
      "name": "dev",
      "description": "This version of the document is for development purposes"
    }
  ],
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/v1/coffees/": {
      "get": {
        "tags": [
          "Coffee"
        ],
        "description": "Gets all coffees",
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/CoffeeArray"
            }
          }
        }
      },
      "post": {
        "tags": [
          "Coffee"
        ],
        "description": "Create a new coffee resource",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Add a coffee",
            "schema": {
              "$ref": "#/definitions/AddCoffee"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Coffee created successfully",
            "schema": {
              "$ref": "#/definitions/Coffee"
            }
          },
          "400": {
            "description": "Invalid payload",
            "schema": {
              "$ref": "#/definitions/FailMessage"
            }
          }
        }
      }
    },
    "/api/v1/coffees/{id}": {
      "get": {
        "tags": [
          "Coffee"
        ],
        "description": "Get coffee by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID of coffee"
          }
        ],
        "responses": {
          "200": {
            "description": "Coffee resource matching ID",
            "schema": {
              "$ref": "#/definitions/Coffee"
            }
          },
          "400": {
            "description": "Invalid ID",
            "schema": {
              "$ref": "#/definitions/FailMessage"
            }
          },
          "404": {
            "description": "Coffee couldnt be found with that ID",
            "schema": {
              "$ref": "#/definitions/FailMessage"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "Coffee": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 1
        },
        "name": {
          "type": "string",
          "example": "Example coffee"
        },
        "details": {
          "type": "string",
          "example": "Example details"
        },
        "price": {
          "type": "number",
          "example": 5.99
        }
      }
    },
    "CoffeeArray": {
      "type": "array",
      "items": {
        "$ref": "#/definitions/Coffee"
      }
    },
    "AddCoffee": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "Example coffee"
        },
        "details": {
          "type": "string",
          "example": "Example details"
        },
        "price": {
          "type": "number",
          "example": 5.99
        }
      },
      "required": [
        "name",
        "details",
        "price"
      ]
    },
    "FailMessage": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "example": "Failure message"
        }
      }
    }
  }
}